requirements.txt
flask
python-dotenv
psycopg2-binary
psycopg2
sqlalchemy
Flask-Migrate